requirements.txt
tensorflow
biopython
pandas
numpy
